requirements.txt
absl-py==0.11.0
deepctr==0.8.2
joblib==1.0.0
numpy==1.19.4
pandas==1.1.5
scikit-learn==0.24.0
sklearn==0.0
tensorflow==2.4.0
